# docs/swagger.yaml
definitions:
  model.Product:
    properties:
      code:
        type: string
      id:
        type: integer
      name:
        type: string
      quantity:
        type: integer
      size:
        type: integer
      warehouse:
        items:
          $ref: '#/definitions/model.ProductWarehouse'
        type: array
    type: object
  model.ProductCreateReq:
    properties:
      code:
        type: string
      name:
        type: string
      quantity:
        type: integer
      size:
        type: integer
      warehouse:
        items:
          type: integer
        type: array
    type: object
  model.ProductWarehouse:
    properties:
      id:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
      warehouse_id:
        type: integer
    type: object
  model.Reservation:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
      warehouse_id:
        type: integer
    type: object
  model.Warehouse:
    properties:
      available:
        type: boolean
      id:
        type: integer
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/model.Product'
        type: array
    type: object
  model.WarehouseCreateReq:
    properties:
      available:
        type: boolean
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /product/:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: List of products
          schema:
            items:
              $ref: '#/definitions/model.Product'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: product
      summary: List Products
      tags:
      - Product
    post:
      consumes:
      - application/json
      parameters:
      - description: Product
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/model.ProductCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            type: product
        "400":
          description: Bad Request
          schema:
            type: product
        "500":
          description: Internal Server Error
          schema:
            type: product
      summary: Create
      tags:
      - Product
  /product/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            type: product
        "400":
          description: Bad Request
          schema:
            type: product
        "500":
          description: Internal Server Error
          schema:
            type: product
      summary: Get Product by ID
      tags:
      - Product
  /product/available/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: warehouse ID
        in: path
        name: warehouseID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of products
          schema:
            items:
              $ref: '#/definitions/model.ProductWarehouse'
            type: array
        "400":
          description: Bad Request
          schema:
            type: product
        "500":
          description: Internal Server Error
          schema:
            type: product
      summary: Get Available Quantity
      tags:
      - Product
  /product/release:
    post:
      consumes:
      - application/json
      parameters:
      - description: Release request
        in: body
        name: ReleaseRequest
        required: true
        schema:
          items:
            $ref: '#/definitions/model.Reservation'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: product
        "500":
          description: Internal Server Error
          schema:
            type: product
      summary: Release Reserved Products
      tags:
      - Product
  /product/reserve:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reservation request
        in: body
        name: Reservation
        required: true
        schema:
          items:
            $ref: '#/definitions/model.Reservation'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: product
        "500":
          description: Internal Server Error
          schema:
            type: product
      summary: Reservation Product
      tags:
      - Product
  /warehouse/:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: List of warehouse
          schema:
            items:
              $ref: '#/definitions/model.Warehouse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: warehouse
      summary: List Warehouses
      tags:
      - Warehouse
    post:
      consumes:
      - application/json
      parameters:
      - description: Warehouse
        in: body
        name: warehouse
        required: true
        schema:
          $ref: '#/definitions/model.WarehouseCreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            type: warehouse
        "500":
          description: Internal Server Error
          schema:
            type: warehouse
      summary: Create Warehouse
      tags:
      - Warehouse
  /warehouse/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Warehouse ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Warehouse details
          schema:
            type: warehouse
        "400":
          description: Bad Request
          schema:
            type: warehouse
        "500":
          description: Internal Server Error
          schema:
            type: warehouse
      summary: Get Warehouse by ID
      tags:
      - Warehouse
swagger: "2.0"